file input and output

+++ 0x15-file_io/1-create_file.c
@@ -5,7 +5,7 @@
 * @s: input
 * Return: string size
 */
-int str_len(char *s)
+ssize_t str_len(char *s)
 {
 	int i;
